+++ client/AdminEvent.js
@@ -1,9 +1,13 @@
 import React, { useState, useEffect } from 'react';
-import { Button, TextField, Card, CardContent, Typography, FormControl, RadioGroup, FormControlLabel, Radio, Paper } from '@mui/material';
+import { Box, Button, TextField, Card, CardContent, Typography, FormControl, RadioGroup, FormControlLabel, Radio, Paper, Chip } from '@mui/material';
 
-export default function AdminEvent({ eventId }) {
+export default function AdminEvent({ event_id }) {
+  const [newBetPrompt, setNewBetPrompt] = useState('');
+  const [newBetType, setNewBetType] = useState(null);
+  const [newBetPoints, setNewBetPoints] = useState(1);
   const [questions, setQuestions] = useState([]);
   const [correctAnswers, setCorrectAnswers] = useState({});
+  const [callForQuestionsNow, setCallForQuestionsNow] = useState(false);
 
 //   const [questions, setQuestions] = useState([
 //     { bet_id: '1', question: 'Is React a library for frontend development?', type: 'true_false' },
@@ -37,18 +41,47 @@ export default function AdminEvent({ eventId }) {
 
 
   useEffect(() => {
-    fetch(`/get-questionnaire/${eventId}`)
+    fetch(`/get-questionnaire/${event_id}`)
       .then(response => response.json())
       .then(data => {
         setQuestions(data.questionnaire);
         let initialCorrectAnswers = {};
         data.questionnaire.forEach(question => {
+          // what are we doing here
           initialCorrectAnswers[question.bet_id] = '';
         });
         setCorrectAnswers(initialCorrectAnswers);
       })
       .catch(error => console.error("Couldn't fetch questions:", error));
-  }, [eventId]);
+  }, [event_id]);
+
+  const updateNewBetPrompt = (e) => {
+    setNewBetPrompt(e.target.value);
+    console.log('newBetPrompt:', e.target.value)
+  };
+
+  const updateNewBetPoints = (e) => {
+    setNewBetPoints(e.target.value);
+    console.log('newBetPoints:', e.target.value)
+  };
+
+  const handleAddBet = async () => {
+    // const addBetRes = await fetch(`/event/new-bet/${event_id}`, {
+      const addBetRes = await fetch(`/event/new-bet/6`, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({
+        type: newBetType,
+        question: newBetPrompt,
+        points: newBetPoints
+      }),
+    })
+    const newBet = await addBetRes.json();
+    console.log('newBetData:', newBet);
+    setQuestions(prev => [...prev, newBet]);
+  };
 
   const handleAnswerChange = (betId, answer) => {
     setCorrectAnswers(prev => ({ ...prev, [betId]: answer }));
@@ -60,7 +93,7 @@ export default function AdminEvent({ eventId }) {
       headers: {
         'Content-Type': 'application/json',
       },
-      body: JSON.stringify({ eventId, correctAnswers }),
+      body: JSON.stringify({ event_id, correctAnswers }),
     })
     .then(response => {
       if (!response.ok) {
@@ -73,8 +106,39 @@ export default function AdminEvent({ eventId }) {
 
   return (
     <div>
-      {questions.map((question) => (
-        <Paper key={question.bet_id} elevation={3} sx={{ marginBottom: 2, width: '60vw', marginLeft: 'auto', marginRight: 'auto' }}>
+      <Box>
+        <Typography variant="h5">
+                Add Bets!
+            </Typography>
+            <Chip label="true or false" 
+            color={newBetType === 'true_false' ? 'primary' : 'default'} 
+            onClick={() => setNewBetType('true_false')} />
+            <Chip label="Player Input" 
+            color={newBetType === 'player_input' ? 'primary' : 'default'} 
+            onClick={() => setNewBetType('player_input')}/>
+        </Box>
+                  <FormControl>
+        {newBetType !== null && 
+            <Box>
+                <TextField label="Question" 
+                onChange={updateNewBetPrompt}
+                />
+                <TextField label="Points" 
+                type="number" 
+                value={newBetPoints} 
+                onChange={updateNewBetPoints}
+                />
+                <Button onClick={handleAddBet} 
+                variant="contained" 
+                color="success" >
+                    Add Bet
+                </Button>
+            </Box>
+      }
+      </FormControl>
+    <Box>
+      {questions.map((question, i) => (
+        <Paper key={i} elevation={3} sx={{ marginBottom: 2, width: '60vw', marginLeft: 'auto', marginRight: 'auto' }}>
           <Card>
             <CardContent>
               <Typography variant="h6">{question.question}</Typography>
@@ -108,6 +172,7 @@ export default function AdminEvent({ eventId }) {
       <Button variant="contained" color="primary" onClick={handleSubmitCorrectAnswers} sx={{ marginTop: 2 }}>
         Submit Correct Answers
       </Button>
+    </Box>
     </div>
   );
 }
